file_backend: obliterate
We can't deploy this and we are not going to support it, so it just
needs to go.
