+++ server/Cargo.toml
@@ -15,7 +15,6 @@ tracing-opentelemetry.workspace = true
 tracing-subscriber.workspace = true
 init-tracing-opentelemetry.workspace = true
 opentelemetry_sdk = { workspace = true, features = ["tokio", "http", "opentelemetry-http", "serde"] }
-xattr.workspace = true
 
 storage = { path = "../storage" }
 tower.workspace = true
@@ -24,7 +23,6 @@ tower-service.workspace = true
 serde.workspace = true
 config.workspace = true
 thiserror.workspace = true
-tempfile.workspace = true
 http-body-util.workspace = true
 tokio-stream.workspace = true
 tokio-util.workspace = true
