+++ storage/Cargo.toml
@@ -7,10 +7,8 @@ edition.workspace = true
 thiserror.workspace = true
 async-trait.workspace = true
 tokio.workspace = true
-rustix.workspace = true
 camino.workspace = true
 pin-project.workspace = true
-xattr.workspace = true
 tracing.workspace = true
 chrono.workspace = true
 sqlx.workspace = true
@@ -19,6 +17,3 @@ futures-core.workspace = true
 # used for three u32 Whence constants >_>
 nix.workspace = true
 assert-impl.workspace = true
-
-[dev-dependencies]
-tempfile.workspace = true
